requirements.txt
Django==1.4.2
PIL==1.1.7
South==0.7.6
boto==2.6.0
dj-database-url==0.2.1
django-appconf==0.5
django-heroku-memcacheify==0.3
django-imagekit==2.0.2
django-pylibmc-sasl==0.2.4
django-social-auth==0.7.10
django-storages==1.1.5
gunicorn==0.16.1
httplib2==0.7.7
oauth2==1.5.211
psycopg2==2.4.5
pylibmc==1.2.3
python-openid==2.2.5
redis==2.7.2
wsgiref==0.1.2
